Screen Scaling Needed to scale speed of player when window resolution changes. Otherwise the consistent pixel per second speed would be inversely proportional to the screams scale. Eg.     32 px width -> cross window with in 10 sec     64 px width -> cross window with in 20 sec
Bug arose when applying scale factor didn't make the speed consistent.
Solution: Fame independence was missing from the player movement calculations. As such, when the resolution decreased, despite the scale factor, the player would move faster as a result of the lessened load on the computer.
Identified cause of inconsistency using a using a desk check on google sheets to confirm that the math was correct.
https://i.imgur.com/RSRaBUN.png